+++ History.md
@@ -1,5 +1,15 @@
 # Changes
 
+## 2.0.0 / 2013-01-20
+
+  - Enforce `reduce` contract on the all `reducibles` by making it part of
+    `reduce` instead of `reducible`.
+  - Anything send after `end` or error in `reducible` is logged as warning
+    and ignored down the flow.
+  - After interruption of reducible via `reduced(state)` return value from
+    consumer must be followed by `end` or error value. If `reducible` sends
+    anything else, warning is logged and `end` of `reducible` is enforced.
+
 ## 1.0.5 / 2012-12-18
 
   - Change implementation detail of [reducible][], such that no exceptions